@@ -15,19 +15,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "~13.3.9",
-    "@angular/animations": "~13.3.11",
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular/animations": "^14.2.11",
     "@angular/cdk": "^13.3.9",
-    "@angular/cli": "~13.3.9",
-    "@angular/common": "~13.3.11",
-    "@angular/compiler": "~13.3.11",
-    "@angular/compiler-cli": "~13.3.11",
-    "@angular/core": "~13.3.11",
-    "@angular/forms": "~13.3.11",
+    "@angular/cli": "^14.2.10",
+    "@angular/common": "^14.2.11",
+    "@angular/compiler": "^14.2.11",
+    "@angular/compiler-cli": "^14.2.11",
+    "@angular/core": "^14.2.11",
+    "@angular/forms": "^14.2.11",
     "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "~13.3.11",
-    "@angular/platform-browser-dynamic": "~13.3.11",
-    "@angular/router": "~13.3.11",
+    "@angular/platform-browser": "^14.2.11",
+    "@angular/platform-browser-dynamic": "^14.2.11",
+    "@angular/router": "^14.2.11",
     "@types/chart.js": "^2.9.29",
     "angular-bootstrap-md": "^13.0.0",
     "bootstrap": "^4.6.0",
